(platforms) Native, fix output port type on Source byte, adding option to print the framerate every five seconds in SDL. Node execution now supports better handling of small tokens.

+++ platform-multicore/src/main/resources/lib/art-native/src/art_Source_byte.c
@@ -145,7 +145,7 @@ static void setParam(AbstractActorInstance *pBase, const char *paramName,
 }
 
 static const PortDescription outputPortDescriptions[] = {{0, "Out",
-                                                                 sizeof(int32_t)}};
+                                                                 sizeof(uint8_t)}};
 
 static const int portRate_0[] = {0};
